+++ plugin/src/main/java/net/thenextlvl/service/controller/hologram/DecentHologramController.java
@@ -7,11 +7,11 @@ import net.kyori.adventure.text.serializer.legacy.LegacyComponentSerializer;
 import net.thenextlvl.service.api.hologram.Hologram;
 import net.thenextlvl.service.api.hologram.HologramController;
 import net.thenextlvl.service.api.hologram.HologramLine;
-import net.thenextlvl.service.model.hologram.DecentBlockHologramLine;
-import net.thenextlvl.service.model.hologram.DecentEntityHologramLine;
-import net.thenextlvl.service.model.hologram.DecentHologram;
-import net.thenextlvl.service.model.hologram.DecentItemHologramLine;
-import net.thenextlvl.service.model.hologram.DecentTextHologramLine;
+import net.thenextlvl.service.model.hologram.decent.DecentBlockHologramLine;
+import net.thenextlvl.service.model.hologram.decent.DecentEntityHologramLine;
+import net.thenextlvl.service.model.hologram.decent.DecentHologram;
+import net.thenextlvl.service.model.hologram.decent.DecentItemHologramLine;
+import net.thenextlvl.service.model.hologram.decent.DecentTextHologramLine;
 import org.bukkit.Location;
 import org.bukkit.World;
 import org.bukkit.block.data.BlockData;
