+++ plugin/src/main/java/net/thenextlvl/service/model/hologram/decent/DecentEntityHologramLine.java
@@ -1,4 +1,4 @@
-package net.thenextlvl.service.model.hologram;
+package net.thenextlvl.service.model.hologram.decent;
 
 import eu.decentsoftware.holograms.api.holograms.HologramLine;
 import net.thenextlvl.service.api.hologram.LineType;
