+++ plugin/src/main/java/net/thenextlvl/service/model/hologram/decent/DecentHologram.java
@@ -1,4 +1,4 @@
-package net.thenextlvl.service.model.hologram;
+package net.thenextlvl.service.model.hologram.decent;
 
 import eu.decentsoftware.holograms.api.DHAPI;
 import net.thenextlvl.service.api.hologram.Hologram;
